\name{phased}

\alias{phased}

\docType{data}

\title{Phased Hapmap data}

\description{Reference data from the hapmap project.}

\usage{data(asw1)}

\format{
  A Large matrix with one row per haplotype and one column per SNP
}

\details{
  There are separate data file for each chromosome for each of the following Hapmap populations: ASW, CEU, CHB, JPT, YRI.
}

\examples{
data(asw1)
data(jpt12)
}

\keyword{datasets}
